Split controller.js file into separate modules
