+++ router.js
@@ -1,16 +1,17 @@
 import Router from "express";
-import Controller from './Controller.js'
+import PatientController from './Controllers/PatientController.js'
+import ResolutionController from './Controllers/ResolutionController.js'
+import QueueController from './Controllers/QueueController.js'
 
 const router = new Router();
 
 
-router.post('/patient', Controller.createPatient);
-router.post('/resolution/get', Controller.getResolution);
-// router.get('/patient/next', [Controller.deleteFirstFromQueue, Controller.getCurrentInQueue]);
-router.get('/patient/next', Controller.nextPatientInQueue);
-router.post('/resolution', Controller.createResolution);
-router.get('/patient/all', Controller.getAllPatients);
-router.delete('/resolution:key', Controller.deleteResolution);
+router.post('/patient', PatientController.createPatient);
+router.post('/resolution/get', ResolutionController.getResolution);
+router.get('/patient/next', QueueController.nextPatientInQueue);
+router.post('/resolution', ResolutionController.createResolution);
+router.get('/patient/all', PatientController.getAllPatients);
+router.delete('/resolution/:key', ResolutionController.deleteResolution);
 
 
 export default router;
